+++ opennmt/utils/beam_search.py
@@ -25,6 +25,8 @@ import tensorflow as tf
 
 from tensorflow.python.util import nest
 
+from tensorflow.python.estimator.util import fn_args
+
 # Assuming EOS_ID is 1
 EOS_ID = 1
 # Default value for INF
@@ -108,7 +110,12 @@ def get_state_shape_invariants(tensor):
 
 
 def _log_prob_from_logits(logits):
-  return logits - tf.reduce_logsumexp(logits, axis=2, keepdims=True)
+  # Silence deprecation warning in TensorFlow 1.5+ by using the renamed argument.
+  if "keepdims" in fn_args(tf.reduce_logsumexp):
+    kwargs = {"keepdims": True}
+  else:
+    kwargs = {"keep_dims": True}
+  return logits - tf.reduce_logsumexp(logits, axis=2, **kwargs)
 
 
 def compute_batch_indices(batch_size, beam_size):